requirements.txt
selenium
pytest
pytest-xdist
allure-pytest
pandas
openpyxl
requests
